Designed Snackbar & seperated all dialog components.
