+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/profile/profile-dashboard/change-password-dialog.component.ts
@@ -1,0 +1,8 @@
+﻿import { Component } from "@angular/core";
+
+@Component({
+  moduleId: module.id,
+  selector: 'change-password-dialog',
+  templateUrl: "change-password-dialog.html"
+})
+export class ChangePasswordDialogComponent { }
