+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/profile/profile-dashboard/profile-dashboard.component.ts
@@ -1,25 +1,19 @@
 ﻿import { Component, OnInit, ViewChild } from "@angular/core";
 import { MdDialog } from '@angular/material';
+import { ChangePasswordDialogComponent } from "./change-password-dialog.component";
 
 @Component({
-    moduleId: module.id,
-    selector: "profile-dashboard",
-    templateUrl: "profile-dashboard.html"
+  moduleId: module.id,
+  selector: "profile-dashboard",
+  templateUrl: "profile-dashboard.html"
 })
 
-export class ProfileDashboardComponent{
+export class ProfileDashboardComponent {
 
-    // Open Change Password Dialog
-    constructor(public dialog: MdDialog) { }
+  // Open Change Password Dialog
+  constructor(public dialog: MdDialog) { }
 
-    changePasswordDialog() {
-        this.dialog.open(ChangePasswordDialogComponent);
-    }
+  changePasswordDialog() {
+    this.dialog.open(ChangePasswordDialogComponent);
+  }
 }
-
-@Component({
-    moduleId: module.id,
-    selector: 'change-password-dialog',
-    templateUrl: "change-password-dialog.html"
-})
-export class ChangePasswordDialogComponent { }
