+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/questions/questions-dashboard/add-category-dialog.component.ts
@@ -1,0 +1,8 @@
+﻿import { Component } from "@angular/core";
+
+@Component({
+    moduleId: module.id,
+    selector: 'add-category-dialog',
+    templateUrl: "add-category-dialog.html"
+})
+export class AddCategoryDialogComponent { }
