+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/questions/questions-dashboard/delete-category-dialog.component.ts
@@ -1,0 +1,8 @@
+﻿import { Component } from "@angular/core";
+
+@Component({
+  moduleId: module.id,
+  selector: 'delete-category-dialog',
+  templateUrl: "delete-category-dialog.html"
+})
+export class DeleteCategoryDialogComponent { }
