+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/questions/questions-dashboard/delete-question-dialog.component.ts
@@ -1,0 +1,8 @@
+﻿import { Component } from "@angular/core";
+
+@Component({
+  moduleId: module.id,
+  selector: 'delete-question-dialog',
+  templateUrl: "delete-question-dialog.html"
+})
+export class DeleteQuestionDialogComponent { }
