+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/questions/questions-dashboard/questions-dashboard.html
@@ -66,7 +66,7 @@
                         <i class="material-icons pull-right" [mdMenuTriggerFor]="quesMoreMenu">more_vert</i>
                         <md-menu #quesMoreMenu="mdMenu" [overlapTrigger]="false" x-position="before">
                             <button md-menu-item>Edit</button>
-                            <button md-menu-item>Delete</button>
+                            <button md-menu-item (click)="deleteQuestionDialog()">Delete</button>
                             <button md-menu-item>Duplicate</button>
                         </md-menu>
                     </div>
@@ -90,7 +90,7 @@
                         <i class="material-icons pull-right" [mdMenuTriggerFor]="quesMoreMenu">more_vert</i>
                         <md-menu #quesMoreMenu="mdMenu" [overlapTrigger]="false" x-position="before">
                             <button md-menu-item>Edit</button>
-                            <button md-menu-item>Delete</button>
+                            <button md-menu-item (click)="deleteQuestionDialog()">Delete</button>
                             <button md-menu-item>Duplicate</button>
                         </md-menu>
                     </div>
@@ -114,7 +114,7 @@
                         <i class="material-icons pull-right" [mdMenuTriggerFor]="quesMoreMenu">more_vert</i>
                         <md-menu #quesMoreMenu="mdMenu" [overlapTrigger]="false" x-position="before">
                             <button md-menu-item>Edit</button>
-                            <button md-menu-item>Delete</button>
+                            <button md-menu-item (click)="deleteQuestionDialog()">Delete</button>
                             <button md-menu-item>Duplicate</button>
                         </md-menu>
                     </div>
