+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/questions/questions.module.ts
@@ -3,7 +3,10 @@ import { SharedModule } from '../shared/shared.module';
 
 import { questionsRouting } from "./questions.routing";
 import { QuestionsComponent } from "./questions.component";
-import { QuestionsDashboardComponent, AddCategoryDialogComponent, DeleteCategoryDialogComponent } from "./questions-dashboard/questions-dashboard.component";
+import { QuestionsDashboardComponent } from "./questions-dashboard/questions-dashboard.component";
+import { AddCategoryDialogComponent } from "./questions-dashboard/add-category-dialog.component";
+import { DeleteCategoryDialogComponent } from "./questions-dashboard/delete-category-dialog.component";
+import { DeleteQuestionDialogComponent } from "./questions-dashboard/delete-question-dialog.component";
 import { QuestionsSingleAnswerComponent } from "./questions-single-answer/questions-single-answer.component";
 import { QuestionsMultipleAnswersComponent } from './questions-multiple-answers/questions-multiple-answers.component';
 import { QuestionsProgrammingComponent } from './questions-programming/questions-programming.component';
@@ -22,11 +25,13 @@ import { CategoryService } from "./categories.service";
         QuestionsSingleAnswerComponent,
         QuestionsMultipleAnswersComponent,
         QuestionsProgrammingComponent,
-        DeleteCategoryDialogComponent
+        DeleteCategoryDialogComponent,
+        DeleteQuestionDialogComponent
     ],
     entryComponents: [
       AddCategoryDialogComponent,
-      DeleteCategoryDialogComponent
+      DeleteCategoryDialogComponent,
+      DeleteQuestionDialogComponent
     ],
     providers: [
         QuestionsService,
