+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/tests/test-settings/test-launch-dialog.component.ts
@@ -1,0 +1,8 @@
+﻿import { Component } from "@angular/core";
+
+@Component({
+    moduleId: module.id,
+    selector: 'test-launch-dialog',
+    templateUrl: "test-launch-dialog.html"
+})
+export class TestLaunchDialogComponent { }
