+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/tests/test-settings/test-settings.component.ts
@@ -1,5 +1,6 @@
 ﻿import { Component, OnInit, ViewChild } from "@angular/core";
 import { MdDialog } from '@angular/material';
+import { TestLaunchDialogComponent } from "./test-launch-dialog.component";
 
 @Component({
     moduleId: module.id,
@@ -18,11 +19,3 @@ export class TestSettingsComponent{
         this.dialog.open(TestLaunchDialogComponent);
     }
 }
-
-@Component({
-    moduleId: module.id,
-    selector: 'test-launch-dialog',
-    templateUrl: "test-launch-dialog.html"
-})
-export class TestLaunchDialogComponent { }
-
