+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/tests/tests-dashboard/delete-test-dialog.component.ts
@@ -1,0 +1,8 @@
+﻿import { Component } from "@angular/core";
+
+@Component({
+  moduleId: module.id,
+  selector: 'delete-test-dialog',
+  templateUrl: "delete-test-dialog.html"
+})
+export class DeleteTestDialogComponent { }
