+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/tests/tests-dashboard/delete-test-dialog.html
@@ -1,0 +1,13 @@
+﻿<div class="dialog">
+    <div class="dialog-title clearfix">
+        <h1 class="pull-left">Delete Confirmation</h1>
+        <button type="button" class="pull-right" md-dialog-close><i class="material-icons">close</i></button>
+    </div>
+    <div class="dialog-body">
+        <h4 class="mb20">Are you sure you want to delete this test?</h4>
+    </div>
+    <div class="dialog-actions">
+        <button type="button" class="btn" md-dialog-close>No</button>
+        <button type="submit" class="btn btn-primary">Yes</button>
+    </div>
+</div>
