+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/tests/tests-dashboard/test-create-dialog.component.ts
@@ -1,0 +1,8 @@
+﻿import { Component } from "@angular/core";
+
+@Component({
+    moduleId: module.id,
+    selector: 'test-create-dialog',
+    templateUrl: "test-create-dialog.html"
+})
+export class TestCreateDialogComponent { }
