+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/tests/tests-dashboard/tests-dashboard.html
@@ -23,7 +23,7 @@
                                 <i class="material-icons pull-right" [mdMenuTriggerFor]="testMenu">more_vert</i>
                                 <md-menu #testMenu="mdMenu" [overlapTrigger]="false" x-position="before">
                                     <a md-menu-item>Edit</a>
-                                    <a md-menu-item>Delete</a>
+                                    <a md-menu-item (click)="deleteTestDialog()">Delete</a>
                                     <a md-menu-item>Duplicate</a>
                                     <a md-menu-item [routerLink]="['settings']">Settings</a>
                                 </md-menu>
@@ -43,4 +43,3 @@
         </div>
     </div>
 </div>
-
