+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/tests/tests.module.ts
@@ -2,8 +2,11 @@
 import { SharedModule } from '../shared/shared.module';
 import { testsRouting } from "./tests.routing";
 import { TestsComponent } from "./tests.component";
-import { TestsDashboardComponent, TestCreateDialogComponent } from "./tests-dashboard/tests-dashboard.component";
-import { TestSettingsComponent, TestLaunchDialogComponent } from "./test-settings/test-settings.component";
+import { TestsDashboardComponent } from "./tests-dashboard/tests-dashboard.component";
+import { TestCreateDialogComponent } from "./tests-dashboard/test-create-dialog.component";
+import { DeleteTestDialogComponent } from "./tests-dashboard/delete-test-dialog.component";
+import { TestSettingsComponent } from "./test-settings/test-settings.component";
+import { TestLaunchDialogComponent } from "./test-settings/test-launch-dialog.component";
 import { TestService } from "./tests.service";
 import { FilterPipe } from "./tests-dashboard/test-dashboard.pipe";
 
@@ -16,6 +19,7 @@ import { FilterPipe } from "./tests-dashboard/test-dashboard.pipe";
         TestsComponent,
         TestsDashboardComponent,
         TestCreateDialogComponent,
+        DeleteTestDialogComponent,
         TestSettingsComponent,
         TestLaunchDialogComponent,
         FilterPipe
@@ -23,6 +27,7 @@ import { FilterPipe } from "./tests-dashboard/test-dashboard.pipe";
     ],
     entryComponents: [
         TestCreateDialogComponent,
+        DeleteTestDialogComponent,
         TestLaunchDialogComponent
     ],
     providers: [
